fix: prod deploy
