+++ src/native/androidProject/app/src/main/res/layout/activity_main.xml
@@ -10,8 +10,7 @@
         android:layout_height="0dp"
         app:layout_constraintTop_toTopOf="parent"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintBottom_toTopOf="@+id/bottom_bar">
+        app:layout_constraintEnd_toEndOf="parent">
 
         <WebView
             android:id="@+id/webview"
